Overlapping Rooms with Edge detection
Removed the graph data structure code. Implemented code that allows rooms to overlap, but honours walls from the existing room.
Not too pretty. Also tiles are only instantiated for the rooms and their walls etc. This drastically reduces the number of game objects required allowing larger maps with larger rooms and / or larger room counts.

Need to lower the runtime of for loops later. Need to tidy up the map. Need to find a way to create passages :P